+++ src/init.cpp
@@ -1435,7 +1435,7 @@ static ChainstateLoadResult InitAndLoadChainstate(
             LogInfo("Block index and chainstate loaded");
             auto mining_args{node::ReadMiningArgs(args)};
             Assert(mining_args); // no error can happen, already checked in AppInitParameterInteraction
-            node.block_template_manager = std::make_unique<node::BlockTemplateManager>(*node.mempool, chainman, std::move(*mining_args));
+            node.block_template_manager = std::make_unique<node::BlockTemplateManager>(*node.mempool, chainman, *node.notifications, std::move(*mining_args));
             node.notifications->setChainstateLoaded(true);
         }
     }
